Avoid using ENABLE_WALLET where not required
