+++ src/init.h
@@ -16,9 +16,6 @@ class CWallet;
 class HTTPRPCRequestProcessor;
 class RPCServer;
 
-class WalletInitInterface;
-extern WalletInitInterface *const g_wallet_init_interface;
-
 namespace boost {
 class thread_group;
 } // namespace boost
